Created Helper class and moved extractNames() there
