+++ src/main/java/training/supportbank/Helper.java
@@ -1,0 +1,19 @@
+package training.supportbank;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+public class Helper {
+    public static Set<String> extractNames(List<Transaction> transactions) {
+        Set<String> accountNames = new HashSet<>();
+
+        for (int i = 0; i < transactions.size(); i++) {
+            for(Transaction transaction : transactions) {
+                accountNames.add(transaction.from);
+                accountNames.add(transaction.to);
+            }
+        }
+        return accountNames;
+    }
+}
